[Refactor] Move isUtil function in colorTemplate to isUtil.ts
